Rename Curve->Spline, use sparse algebra when fitting
