+++ examples/spline_reparameterize.cpp
@@ -27,9 +27,10 @@
  * @file bspline.cpp Curve example.
  */
 
-#include "smooth/spline/curve.hpp"
 #include "smooth/se2.hpp"
-#include "smooth/spline/fit_curve.hpp"
+#include "smooth/spline/fit_spline.hpp"
+#include "smooth/spline/reparameterize.hpp"
+#include "smooth/spline/spline.hpp"
 
 #ifdef ENABLE_PLOTTING
 #include <matplot/matplot.h>
@@ -40,7 +41,7 @@
  */
 int main(int, char const **)
 {
-  using Curve = smooth::CubicCurve<smooth::SE2d>;
+  using Curve = smooth::CubicSpline<smooth::SE2d>;
   std::srand(100);
 
   std::vector<double> tt{1, 2, 3, 4, 5, 6};
@@ -53,7 +54,7 @@ int main(int, char const **)
     smooth::SE2d::Random(),
   };
 
-  Curve c = smooth::fit_curve(tt, gg, smooth::FixedDerCubic<smooth::SE2d, 1>{});
+  Curve c = smooth::fit_spline(tt, gg, smooth::FixedDerCubic<smooth::SE2d, 1>{});
   c += Curve::ConstantVelocity(Eigen::Vector3d(1.2, 0, 0), 5);
   c += Curve::ConstantVelocity(Eigen::Vector3d(1, 0, 1), 2);
   c += Curve::ConstantVelocity(Eigen::Vector3d(0, 0, 0), 2);
@@ -61,7 +62,7 @@ int main(int, char const **)
 
   Eigen::Vector3d vmax(1, 1, 1), amax(1, 1, 1);
 
-  auto sfun = smooth::reparameterize_curve(c, -vmax, vmax, -amax, amax, 1, 0);
+  auto sfun = smooth::reparameterize_spline(c, -vmax, vmax, -amax, amax, 1, 0);
 
   std::vector<double> tvec, svec, vvec, avec;
   std::vector<double> vx, vy, w, ax, ay, dw;
