+++ include/smooth/spline/fit_spline.hpp
@@ -23,18 +23,18 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef SMOOTH__SPLINE__FIT_CURVE_HPP_
-#define SMOOTH__SPLINE__FIT_CURVE_HPP_
+#ifndef SMOOTH__SPLINE__FIT_SPLINE_HPP_
+#define SMOOTH__SPLINE__FIT_SPLINE_HPP_
 
-#include <Eigen/Cholesky>
-#include <Eigen/Core>
-#include <Eigen/LU>
+#include <Eigen/Sparse>
+#include <Eigen/SparseCholesky>
+#include <Eigen/SparseLU>
 
 #include <cassert>
 #include <ranges>
 
 #include "basis.hpp"
-#include "curve.hpp"
+#include "spline.hpp"
 
 namespace smooth {
 
@@ -187,7 +187,7 @@ auto ss_project(const SS & ss, int k)
  */
 template<SplineSpec SS, std::ranges::range Rt, std::ranges::range Rx>
   requires(std::is_same_v<std::ranges::range_value_t<Rt>, std::ranges::range_value_t<Rx>>)
-Eigen::VectorXd fit_poly_1d(const Rt & dt_r, const Rx & dx_r, const SS & ss = SS{})
+Eigen::VectorXd fit_spline_1d(const Rt & dt_r, const Rx & dx_r, const SS & ss = SS{})
 {
   const std::size_t N = std::min(std::ranges::size(dt_r), std::ranges::size(dx_r));
 
@@ -226,16 +226,18 @@ Eigen::VectorXd fit_poly_1d(const Rt & dt_r, const Rx & dx_r, const SS & ss = SS
 
   // CONSTRAINT MATRICES A, b
 
-  Eigen::MatrixXd A = Eigen::MatrixXd::Zero(N_eq, N_coef);
+  Eigen::SparseMatrix<double> A(N_eq, N_coef);
   Eigen::VectorXd b = Eigen::VectorXd::Zero(N_eq);
+
+  // TODO pre-allocate H
 
   // current inequality counter
   std::size_t M = 0;
 
   // curve beg constraints
   for (auto i = 0u; i < ss.LeftDeg.size(); ++i) {
-    A.row(M).segment(0, K + 1) = U0tB.row(ss.LeftDeg[i]);
-    b(M++)                     = ss.left_values[i].x();
+    for (int j = 0u; j < K + 1; ++j) { A.insert(M, j) = U0tB(ss.LeftDeg[i], j); }
+    b(M++) = ss.left_values[i].x();
   }
 
   auto it_dt = std::ranges::begin(dt_r);
@@ -243,36 +245,42 @@ Eigen::VectorXd fit_poly_1d(const Rt & dt_r, const Rx & dx_r, const SS & ss = SS
 
   for (auto i = 0u; i < N; ++i, ++it_dt, ++it_dx) {
     // interval beg + end constraint
-    A(M, i * (K + 1)) = 1;
-    b(M++)            = 0;
+    A.insert(M, i * (K + 1)) = 1;
+    b(M++)                   = 0;
 
     if (SS::InnCnt >= 0) {
-      A(M, i * (K + 1) + K) = 1;
-      b(M++)                = *it_dx;
+      A.insert(M, i * (K + 1) + K) = 1;
+      b(M++)                       = *it_dx;
     }
 
     // inner derivative continuity constraints
     if (i + 1 < N) {
       for (auto d = 1; d <= SS::InnCnt; ++d) {
-        A.row(M).segment(i * (K + 1), K + 1)       = U1tB.row(d) / std::pow(*it_dt, d);
-        A.row(M).segment((i + 1) * (K + 1), K + 1) = -U0tB.row(d) / std::pow(*(it_dt + 1), d);
-        b(M++)                                     = 0;
+        for (auto j = 0; j < K + 1; ++j) {
+          A.insert(M, i * (K + 1) + j) = U1tB(d, j) / std::pow(*it_dt, d);
+        }
+        for (auto j = 0; j < K + 1; ++j) {
+          A.insert(M, (i + 1) * (K + 1) + j) = -U0tB(d, j) / std::pow(*(it_dt + 1), d);
+        }
+        b(M++) = 0;
       }
     }
   }
 
   // curve end derivative zero constraints
   for (auto i = 0u; i < ss.RghtDeg.size(); ++i) {
-    A.row(M).segment((K + 1) * (N - 1), K + 1) = U1tB.row(ss.RghtDeg[i]);
-    b(M++)                                     = ss.rght_values[i].x();
+    for (int j = 0u; j < K + 1; ++j) {
+      A.insert(M, (K + 1) * (N - 1) + j) = U1tB(ss.RghtDeg[i], j);
+    }
+    b(M++) = ss.rght_values[i].x();
   }
 
   if constexpr (SS::OptDeg < 0) {
     // No optimization, solve directly
     assert(N_eq == N_coef);
-    return A.lu().solve(b);
+    Eigen::SparseLU<decltype(A)> lu(A);
+    return lu.solve(b);
   } else {
-
     static_assert(K >= SS::OptDeg, "K >= OptDeg");
 
     // COST MATRIX P
@@ -298,44 +306,56 @@ Eigen::VectorXd fit_poly_1d(const Rt & dt_r, const Rx & dx_r, const SS & ss = SS
     //   [Q A'; A 0] [x; l] =   [0; b]
     // via LDLt factorization
 
-    Eigen::MatrixXd H(N_eq + N_coef, N_eq + N_coef);
-    H.topLeftCorner(N_coef, N_coef) = Eigen::VectorXd::Constant(N_coef, 1e-6).asDiagonal();
+    Eigen::SparseMatrix<double> H(N_coef + N_eq, N_coef + N_eq);
+
+    // TODO pre-allocate H
 
     for (auto i = 0u; const auto & dt : dt_r | std::views::take(int64_t(N))) {
-      H.block(i * (K + 1), i * (K + 1), K + 1, K + 1) += std::pow(dt, 1 - 2 * int(D)) * P;
+      const double fac = std::pow(dt, 1 - 2 * int(D));
+      for (auto ki = 0u; ki != K + 1; ++ki) {
+        for (auto kj = 0u; kj != K + 1; ++kj) {
+          H.insert(i * (K + 1) + ki, i * (K + 1) + kj) = (ki == kj ? 1e-6 : 0.) + fac * P(ki, kj);
+        }
+      }
       ++i;
     }
-    H.bottomLeftCorner(N_eq, N_coef) = A;
-    H.bottomRightCorner(N_eq, N_eq).setZero();
+
+    using AIter = typename Eigen::SparseMatrix<double, Eigen::ColMajor>::InnerIterator;
+
+    for (auto col = 0u; col != N_coef; ++col) {
+      for (AIter it(A, col); it; ++it) { H.insert(N_coef + it.index(), col) = it.value(); }
+    }
+
+    // H.bottomLeftCorner(N_eq, N_coef) = A;
 
     Eigen::VectorXd rhs(N_coef + N_eq);
     rhs.head(N_coef).setZero();
     rhs.tail(N_eq) = b;
 
-    const Eigen::VectorXd sol = H.ldlt().solve(rhs);
+    Eigen::SimplicialLDLT<decltype(H), Eigen::Lower> ldlt(H);
+    const Eigen::VectorXd sol = ldlt.solve(rhs);
 
     return sol.head(N_coef);
   }
 }
 
 /**
- * @brief Fit a Curve to given points.
+ * @brief Fit a Spline to given points.
  *
  * @tparam G LieGroup
- * @tparam K Curve degree
+ * @tparam K Spline degree
  * @param ts range of times
  * @param gs range of values
  * @return curve c s.t. c(t_i) = g_i for (t_i, g_i) \in zip(ts, gs)
  */
 template<std::ranges::range Rt, std::ranges::range Rg, SplineSpec SS>
-auto fit_curve(const Rt & ts, const Rg & gs, const SS & ss)
+auto fit_spline(const Rt & ts, const Rg & gs, const SS & ss)
 {
   using namespace std::views;
 
   const auto N     = std::min(std::ranges::size(ts), std::ranges::size(gs));
-
   constexpr auto K = SS::Degree;
-  using G = typename extract<SS>::group;
+  using G          = typename extract<SS>::group;
 
   assert(N >= 2);
 
@@ -351,13 +371,13 @@ auto fit_curve(const Rt & ts, const Rg & gs, const SS & ss)
 
   for (auto k = 0u; k < Dof<G>; ++k) {
     const auto ss_proj = ss_project(ss, k);
-    V.row(k) = fit_poly_1d(dts, dgs | transform([k](const auto & v) { return v(k); }), ss_proj);
+    V.row(k) = fit_spline_1d(dts, dgs | transform([k](const auto & v) { return v(k); }), ss_proj);
   }
 
-  Curve<K, G> ret;
+  Spline<K, G> ret;
   for (auto i = 0u; const auto & g : gs) {
     if (i + 1 < N) {
-      ret.concat_global(Curve<K, G>(dts[i], V.template block<Dof<G>, K + 1>(0, i * (K + 1)), g));
+      ret.concat_global(Spline<K, G>(dts[i], V.template block<Dof<G>, K + 1>(0, i * (K + 1)), g));
     } else {
       ret.concat_global(g);
     }
@@ -368,4 +388,4 @@ auto fit_curve(const Rt & ts, const Rg & gs, const SS & ss)
 
 }  // namespace smooth
 
-#endif  // SMOOTH__SPLINE__FIT_CURVE_HPP_
+#endif  // SMOOTH__SPLINE__FIT_SPLINE_HPP_
